+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## [1.0.1] - 2020-7-2
+## [1.0.2] 2020-7-3
+
+-   Modify the method addBloack to insertBlock.
+-   New insertText method to insert text based on cursor position.
+
+## [1.0.1] 2020-7-2
 
 -   Support Highlighting
 -   Support @someone